hotfix: everything

+++ controllers/bids.controller.js
@@ -1,7 +1,7 @@
-const Bid = require('../models/bid.model');
-const httpStatusText = require('../utils/httpStatusText');
-const asyncWrapper = require('../middlewares/asyncWrapper');
-const appError = require('../utils/appError');
+const Bid = require('../models/bid.model')
+const httpStatusText = require('../utils/httpStatusText')
+const asyncWrapper = require('../middlewares/asyncWrapper')
+const appError = require('../utils/appError')
 
 /**
  * Retrieves all bids for a specific auction.
@@ -10,12 +10,19 @@ const appError = require('../utils/appError');
  * @param {Function} next - The next middleware function.
  * @returns {Promise<void>}
  */
-const getAllBids = asyncWrapper(
-    async (req, res, next) => {
-        const bids = await Bid.find({ auctionId: req.params.auction_Id }, { "__v": false });
-        res.json({ status: httpStatusText.SUCCESS, data: { bids }, error: null });
-    }
-);
+const getAllBids = asyncWrapper(async (req, res, next) => {
+    const limit = req.query.limit || 6
+    const page = req.query.page || 1
+    const skip = (page - 1) * limit
+    const bids = await Bid.find(
+        { auctionId: req.params.auctionId },
+        { __v: false }
+    )
+        .limit(limit)
+        .skip(skip)
+        .populate('buyerId')
+    res.json({ status: httpStatusText.SUCCESS, data: { bids }, error: null })
+})
 
 /**
  * Retrieves a bid by its ID.
@@ -24,16 +31,14 @@ const getAllBids = asyncWrapper(
  * @param {Function} next - The next middleware function.
  * @returns {Promise<void>}
  */
-const getBid = asyncWrapper(
-    async (req, res, next) => {
-        const bid = await Bid.findById(req.params.bidId);
-        if (!bid) {
-            const error = appError.create('Bid not found', 404, httpStatusText.FAIL);
-            return next(error);
-        }
-        res.json({ status: httpStatusText.SUCCESS, data: { bid }, error: null });
+const getBid = asyncWrapper(async (req, res, next) => {
+    const bid = await Bid.findById(req.params.bidId)
+    if (!bid) {
+        const error = appError.create('Bid not found', 404, httpStatusText.FAIL)
+        return next(error)
     }
-);
+    res.json({ status: httpStatusText.SUCCESS, data: { bid }, error: null })
+})
 
 /**
  * Creates a new bid.
@@ -41,17 +46,19 @@ const getBid = asyncWrapper(
  * @param {Object} res - The response object.
  * @returns {Promise<void>}
  */
-const createBid = asyncWrapper(
-    async (req, res) => {
-        const newBid = new Bid({
-            auctionId: req.body.auctionId,
-            price: req.body.price,
-            buyerId: req.decodedToken.id
-        });
-        await newBid.save();
-        res.status(201).json({ status: httpStatusText.SUCCESS, data: { newBid }, error: null });
-    }
-);
+const createBid = asyncWrapper(async (req, res) => {
+    const newBid = new Bid({
+        auctionId: req.body.auctionId,
+        price: req.body.price,
+        buyerId: req.decodedToken.id,
+    })
+    await newBid.save()
+    res.status(201).json({
+        status: httpStatusText.SUCCESS,
+        data: { newBid },
+        error: null,
+    })
+})
 
 /**
  * Deletes a bid by its ID.
@@ -60,17 +67,32 @@ const createBid = asyncWrapper(
  * @param {Function} next - The next middleware function.
  * @returns {Promise<void>}
  */
-const deleteBid = asyncWrapper(
-    async (req, res, next) => {
-        const bidId = req.params.bidId;
-        await Bid.deleteOne({ _id: bidId });
-        res.status(200).json({ status: httpStatusText.SUCCESS, data: null, error: null });
+const deleteBid = asyncWrapper(async (req, res, next) => {
+    const bidId = req.params.bidId
+    const bid = await Bid.findById(bidId)
+    if (!bid) {
+        const error = appError.create('Bid not found', 404, httpStatusText.FAIL)
+        return next(error)
     }
-);
+    if (bid.buyerId != req.decodedToken.id) {
+        const error = appError.create(
+            'User not authorized',
+            401,
+            httpStatusText.FAIL
+        )
+        return next(error)
+    }
+    await Bid.deleteOne({ _id: bidId })
+    res.status(200).json({
+        status: httpStatusText.SUCCESS,
+        data: null,
+        error: null,
+    })
+})
 
 module.exports = {
     getAllBids,
     getBid,
     createBid,
-    deleteBid
-};
+    deleteBid,
+}
