+++ controllers/messages.controller.js
@@ -1,9 +1,8 @@
-const { validationResult } = require('express-validator');
-const Message = require('../models/message.model');
-const httpStatusText = require('../utils/httpStatusText');
-const asyncWrapper = require('../middlewares/asyncWrapper');
-const appError = require('../utils/appError');
-
+const Message = require('../models/message.model')
+const httpStatusText = require('../utils/httpStatusText')
+const asyncWrapper = require('../middlewares/asyncWrapper')
+const appError = require('../utils/appError')
+const ChatRoom = require('../models/chatRoom.model')
 
 /**
  * Retrieves all messages for a specific chat room.
@@ -12,30 +11,47 @@ const appError = require('../utils/appError');
  * @param {Function} next - The next middleware function.
  * @returns {Promise<void>}
  */
-const getAllMessages = asyncWrapper(
-    async (req, res, next) => {
-        const messages = await Message.find({ chatRoomId: req.params.chatRoom_Id }, { "__v": false });
-        res.json({ status: httpStatusText.SUCCESS, data: { messages }, error: null });
+const getAllMessages = asyncWrapper(async (req, res, next) => {
+    const limit = req.query.limit || 6
+    const page = req.query.page || 1
+    const skip = (page - 1) * limit
+    const chatRoomId = req.params.chatRoomId
+    const chatRoom = await ChatRoom.findById(chatRoomId)
+    if (!chatRoom) {
+        const error = appError.create(
+            'ChatRoom not found',
+            404,
+            httpStatusText.FAIL
+        )
+        return next(error)
     }
-);
+    if (
+        !(
+            chatRoom.user1 == req.decodedToken.id ||
+            chatRoom.user2 == req.decodedToken.id
+        )
+    ) {
+        const error = appError.create(
+            'User not authorized',
+            401,
+            httpStatusText.FAIL
+        )
+        return next(error)
+    }
+    const messages = await Message.find(
+        { chatRoomId: chatRoomId },
+        { __v: false }
+    )
+        .limit(limit)
+        .skip(skip)
+        .populate('senderId')
 
-/**
- * Retrieves a message by its ID.
- * @param {Object} req - The request object.
- * @param {Object} res - The response object.
- * @param {Function} next - The next middleware function.
- * @returns {Promise<void>}
- */
-const getMessage = asyncWrapper(
-    async (req, res, next) => {
-        const message = await Message.findOne({ _id: req.params.message_Id });
-        if (!message) {
-            const error = appError.create('Message not found', 404, httpStatusText.FAIL);
-            return next(error);
-        }
-        res.json({ status: httpStatusText.SUCCESS, data: { message }, error: null });
-    }
-);
+    res.json({
+        status: httpStatusText.SUCCESS,
+        data: { messages },
+        error: null,
+    })
+})
 
 /**
  * Creates a new message.
@@ -44,32 +60,28 @@ const getMessage = asyncWrapper(
  * @param {Function} next - The next middleware function.
  * @returns {Promise<void>}
  */
-const createMessage = asyncWrapper(
-    async (req, res, next) => {
-        const newMessage = new Message({
-            chatRoomId: req.body.chatRoomId,
-            content: req.body.content,
-            senderId: req.decodedToken.id,
-            createdAt: req.body.createdAt
-        });
-        await newMessage.save();
-        res.status(201).json({ status: httpStatusText.SUCCESS, data: { newMessage }, error: null });
+const createMessage = asyncWrapper(async (req, res, next) => {
+    const chatRoom = await ChatRoom.findById(req.body.chatRoomId)
+    if (!chatRoom) {
+        const error = appError.create(
+            'ChatRoom not found',
+            404,
+            httpStatusText.FAIL
+        )
+        return next(error)
     }
-);
-
-/**
- * Updates a message.
- * @param {Object} req - The request object.
- * @param {Object} res - The response object.
- * @param {Function} next - The next middleware function.
- * @returns {Promise<void>}
- */
-const updateMessage = asyncWrapper(
-    async (req, res, next) => {
-        await Message.updateOne({ _id: req.params.message_Id }, { $set: { ...req.body } });
-        return res.status(200).json({ status: httpStatusText.SUCCESS, data: null, error: null });
-    }
-);
+    const newMessage = new Message({
+        chatRoomId: req.body.chatRoomId,
+        content: req.body.content,
+        senderId: req.decodedToken.id,
+    })
+    await newMessage.save()
+    res.status(201).json({
+        status: httpStatusText.SUCCESS,
+        data: { newMessage },
+        error: null,
+    })
+})
 
 /**
  * Deletes a message by its ID.
@@ -78,17 +90,35 @@ const updateMessage = asyncWrapper(
  * @param {Function} next - The next middleware function.
  * @returns {Promise<void>}
  */
-const deleteMessage = asyncWrapper(
-    async (req, res, next) => {
-        await Message.deleteOne({ _id: req.params.message_Id });
-        res.status(200).json({ status: httpStatusText.SUCCESS, data: null, error: null });
+const deleteMessage = asyncWrapper(async (req, res, next) => {
+    const messageId = req.params.messageId
+    const message = await Message.findById(messageId)
+    if (!message) {
+        const error = appError.create(
+            'Message not found',
+            404,
+            httpStatusText.FAIL
+        )
+        return next(error)
     }
-);
+    if (message.senderId != req.decodedToken.id) {
+        const error = appError.create(
+            'User not authorized',
+            401,
+            httpStatusText.FAIL
+        )
+        return next(error)
+    }
+    await Message.deleteOne({ _id: req.params.message_Id })
+    res.status(200).json({
+        status: httpStatusText.SUCCESS,
+        data: null,
+        error: null,
+    })
+})
 
 module.exports = {
     getAllMessages,
-    getMessage,
     createMessage,
-    updateMessage,
-    deleteMessage
-};
+    deleteMessage,
+}
